docker version update

+++ entry_script/entry_script.sh
@@ -5,18 +5,21 @@
 #    note that these changes are ephemeral and will only last while the docker container is being run
 echo $PWD
 
-ls -lrt external/*
-
 echo "User prompting? Changes saved?"
 
-cp external/input/* input/
-cp external/input_templates/* input_templates/
-cp external/docs/* docs/
-cp external/python/* python/
-cp external/script/* script/
+ls -lrt external/*
+
+# copy the users versions of the codes (if any) to the temp dir inside the container
+[ -f external/input/* ] && cp external/input/* input/
+[ -f external/input_templates/* ] && cp external/input_templates/* input_templates/
+[ -f external/docs/* ] && cp external/docs/* docs/
+[ -f external/python/* ] && cp external/python/* python/
+[ -f external/script/* ] && cp external/script/* script/
+[ -f external/output/* ] && cp external/output/* output/
+
 # 2: copy back to the mapped directories. Thus they will have the most up to date versions AND any gaps filled in with default files stored in the docker image
 # only done in jupyter mode - note that this could cause unexpected behaviour if user was using incorrect dir to update and use script mode?
-
+ls -lrt input/*
 
 # Entry point script for MECSim docker. 
 
@@ -27,11 +30,15 @@ if [ "$1" == "--script" ]; then
   chmod +x script/*.sh
   
   # convert all ipynb to py
-  jupyter nbconvert --to python --template=python.tpl python/*
+  jupyter nbconvert --to python --template=python.tpl python/*.ipynb
 
   ./script/run_mecsim_script.sh
 elif [ "$1" == "--single" ]; then
   echo "Running a single MECSim experiment:"
+  # 2: copy back to the mapped directories (from above) 
+  cp input/* external/input/
+  [ -f output/* ] && cp output/* external/output/
+
   # copy in input file
   cp input/Master.inp ./
   ./MECSim.exe 2> err.txt
@@ -47,9 +54,10 @@ elif [ "$1" == "--jupyter" ]; then
   cp docs/* external/docs/
   cp python/* external/python/
   cp script/* external/script/
+  [ -f output/* ] && cp output/* external/output/
 
   # convert all ipynb to py
-  jupyter nbconvert --to python --template=python.tpl python/*
+  jupyter nbconvert --to python --template=python.tpl python/*.ipynb
   
   dos2unix script/*.sh
   chmod +x script/*.sh
@@ -62,9 +70,13 @@ elif [ "$1" == "--update" ]; then
   cp docs/* external/docs/
   cp python/* external/python/
   cp script/* external/script/
+  [ -f output/* ] && cp output/* external/output/
 
+  # these directories will always have contents - even if just the defaults - backup regardless
   backup_dirs='python
-  input_templates'
+  input_templates
+  docs
+  script'
   for this_dir in $backup_dirs;
   do
     cd $this_dir
@@ -82,8 +94,3 @@ else
   echo " --jupyter  : setup a script using jupyter notebooks via browser"
   echo " --update   : download the latest python scripts and notebooks from the github repository ( https://github.com/garethkennedy/MECSim_Analytics/tree/master/python ). CAUTION this will overwrite any notebooks with the same names in your local python directory.  As a failsafe the existing contents of python/ are copied to python/backup"
 fi
-
-#Try exposing volume with :z option in the end.
-#
-#volumes:
-#    - host_folder:docker_folder:z
